Implement `BenMenu` extension of `Menu`, moving all port-specific code to `BenMenu`. `MenuTypes.h` and `Menu.h` should now be safe to include in multiple places. Window backend dropdown removed from menubar (maybe temporarily?).
